more caching traits and defaults for genericism
