+++ mentat/src/cache/mod.rs
@@ -1,5 +1,11 @@
-mod cached;
-pub use cached::*;
-
 mod cache_trait;
 pub use cache_trait::*;
+
+mod cache_inner_trait;
+pub use cache_inner_trait::*;
+
+mod default_cache;
+pub use default_cache::*;
+
+mod default_inner_cache;
+pub use default_inner_cache::*;
